factory function Reading
