+++ controllers/reviewController.js
@@ -1,21 +1,5 @@
 let Review = require('./../models/reviewModel');
-let catchAsync = require('./../utils/catchAsync');
 let factory = require('./handlerfactoryController');
-
-exports.getAllReviews = catchAsync(async (req, res, next) => {
-  let filter = {};
-  if (req.param.tourId) filter = { tour: req.params.tourId };
-
-  let reviews = await Review.find(filter);
-
-  res.status(200).json({
-    status: 'ok',
-    results: reviews.length,
-    data: {
-      reviews
-    }
-  });
-});
 
 exports.setTourUserIds = (req, res, next) => {
   // Allow nested routes
@@ -25,5 +9,7 @@ exports.setTourUserIds = (req, res, next) => {
 }
 
 exports.createReview = factory.createOne(Review);
+exports.getAllReviews = factory.getAll(Review);
+exports.getReview = factory.getOne(Review);
 exports.updateReview = factory.updateOne(Review);
 exports.deleteReview = factory.deleteOne(Review);
