+++ controllers/userController.js
@@ -12,19 +12,6 @@ let filterObj = (obj, ...allowedFields) => {
   return newObj;
 }
 
-exports.getAllUsers = catchAsync(async (req, res, next) => {
-  let users = await User.find();
-
-  // Send response
-  res.status(200).json({
-      status: "ok",
-      length: users.length,
-      data: {
-        users
-      }
-  });
-});
-
 exports.updateMe = catchAsync(async (req, res, next) => {
   // Create error when user try to change password
   if (req.body.password || req.body.passwordConfirm) {
@@ -60,13 +47,14 @@ exports.deleteMe = catchAsync(async (req, res, next) => {
   });
 });
 
-exports.getUser = (req, res) => {
+exports.createUser = (req, res) => {
   res.status(500).json({
     status: 'error',
-    message: 'Not define'
+    message: 'This Route is Not define! Please use /signup instead.'
   });
 }
 
-exports.createUser = factory.createOne(User);
+exports.getAllUsers = factory.getAll(User);
+exports.getUser = factory.getOne(User);
 exports.updateUser = factory.updateOne(User); // Don't update password with this
 exports.deleteUser = factory.deleteOne(User);
